chore(refactor): separate nvim-treesitter and vim.treesitter related modules
## Details

The `ts` module now only stores logic that interacts directly with
`nvim-treesitter` and its APIs.

The `scopes` method has been moved directly to `selection` module, the
only place it is needed.

The `has` method has been moved to the `util` module and is now called
`has_query`.

This will be used to make most of the logic between `ensure_installed`
and `auto_install` mostly shared in a follow up.