feat: avoid installing the same language more than once
## Details

With auto install enabled we will attempt to install parsers for every
buffer that does not currently have one. With plugin owned buffers (like
oil) a parser will never be found, but we'll still attempt this every
time. To avoid this store a cache of languages we have attempted to
install for the current session, in the event that we have tried to
install the language and no parser exists it likely means there just
isn't one to install, so we skip it.