fix: force visual mode rather than re-using previous
## Details

PR: https://github.com/MeanderingProgrammer/treesitter-modules.nvim/pull/8

Effectively reverts: https://github.com/MeanderingProgrammer/treesitter-modules.nvim/commit/4cb6ee9afdb89ad90532380717cefe3acfa0964e

A side effect of using `gv` to select a range of text is that it re-uses
the previous visual mode. This conflicts with the purpose of incremental
selection which is to select as specific range of text regardless of
what the previous visual mode happened to be. It makes sense in context
of LSP selections, less for incremental, but that's up to opinion. In
this case we'll keep the original behavior from `nvim-treesitter`.

Co-authored-by: PEMessage <1165739182@qq.com>